[ConstraintSystem] NFC: Move new diagnostics into separate files - Diagnostics.(h|cpp)
