+++ lib/Sema/CMakeLists.txt
@@ -11,6 +11,7 @@ add_swift_library(swiftSema STATIC
   CSRanking.cpp
   CSSimplify.cpp
   CSSolver.cpp
+  CSDiagnostics.cpp
   CalleeCandidateInfo.cpp
   CodeSynthesis.cpp
   Constraint.cpp
